% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_relationships.R
\name{format_relationships}
\alias{format_relationships}
\title{Format Relationships as R Code}
\usage{
format_relationships(relationships, object_name = "relationships")
}
\arguments{
\item{relationships}{A relationships list, usually created by
[suggest_relationships()].}

\item{object_name}{Name to use on the left-hand side of the generated code.}
}
\value{
A character string containing R code.
}
\description{
Format Relationships as R Code
}
