% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_erd.R
\name{is_valid_erd}
\alias{is_valid_erd}
\title{Test Whether an ERD Is Valid}
\usage{
is_valid_erd(erd_object, check_data = TRUE, strict = FALSE)
}
\arguments{
\item{erd_object}{An object created by [create_erd()].}

\item{check_data}{Logical. Check observed cardinalities if `TRUE`.}

\item{strict}{Logical. Treat warnings as invalid if `TRUE`.}
}
\value{
A single logical value.
}
\description{
Test Whether an ERD Is Valid
}
