% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_erd.R
\name{create_erd}
\alias{create_erd}
\title{Create ERD Object}
\usage{
create_erd(df_list, relationships)
}
\arguments{
\item{df_list}{A named list of data frames, where each data frame represents
an entity in the ERD. The names of the list elements correspond to the
names of the entities.}

\item{relationships}{A nested named list describing the relationships between
entities. The top-level names in this list should correspond to the names
in \code{df_list}. Each element of this list is itself a list, describing
relationships that the corresponding entity has with other entities. The list of acceptable values is specified in "Details."}
}
\value{
An ERD object
}
\description{
This function serves as a constructor for an Entity-Relationship Diagram
(ERD) object. This object encapsulates both the data frames representing the
entities and the relationships between these entities. The function takes as
its arguments a list of data frames and a list of relationships and returns a
list object of class "ERD".
}
\details{
Possible values in each relationship element of the list include:
\itemize{
 \item{ "||" }{ which indicates one and only one}
 \item{">|" }{ which indicates one or more (left table)}
 \item{"|<" }{ which indicates one or more (right table)}
 \item{">0" }{ which indicates zero or more (left table)}
 \item{"0<" }{ which indicates zero or more (right table)}
 \item{"|0" }{ which indicates zero or one (left table)}
 \item{"0|" }{ which indicates zero or one (right table)}
}
}
\examples{
NULL
}
