% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_erd.R
\name{render_erd}
\alias{render_erd}
\title{Render ERD}
\usage{
render_erd(erd_object, label_distance = 2.5, label_angle = 45, n = 10)
}
\arguments{
\item{erd_object}{An object of class "ERD", generated using the `create_erd`
function. This object encapsulates the data frames representing the
entities and the relationships between these entities.}

\item{label_distance}{A numeric value that specifies the distance between the
edge labels and the lines connecting the nodes. The default value is 2.5.}

\item{label_angle}{A numeric value that specifies the angle at which the edge
labels are displayed. The default value is 45 degrees.}

\item{n}{The maximum number of rows in each table. The tables will add
columns to show all of the variables in the tables such that there are only
`n` rows.}
}
\value{
A `DiagrammeR` graph object representing the ERD.
}
\description{
Render ERD
}
\examples{
NULL
}
